requirements.txt
Django>=4.2,<6
pyjwt==1.6.1
